Change file structure to mimic a package in python. Added tests for Curridata class specifically for its ability as a generator
